# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: 99Placement Recruitment OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Companies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}/contacts:
    get:
      operationId: listCompanyContacts
      tags: [companies]
      summary: List contacts for a company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyContact"
    post:
      operationId: createCompanyContact
      tags: [companies]
      summary: Add contact to company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyContactRequest"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyContact"

  # ── REQUIREMENTS ────────────────────────────────────────────────────────────
...
  /client-interviews:
    get:
      operationId: listClientInterviews
      tags: [clientInterviews]
      summary: List client interviews
      security:
        - bearerAuth: []
      parameters:
        - name: pipelineId
          in: query
          schema:
            type: string
        - name: verdict
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Client interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientInterviewWithCandidate"
    post:
      operationId: createClientInterview
      tags: [clientInterviews]
      summary: Schedule a client interview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientInterviewRequest"
      responses:
        "201":
          description: Scheduled interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInterviewWithCandidate"
  /client-interviews/{id}:
    get:
      operationId: getClientInterview
      tags: [clientInterviews]
      summary: Get interview with feedback
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInterviewWithCandidate"
  /client-interviews/{id}/feedback:
    post:
      operationId: submitClientInterviewFeedback
      tags: [clientInterviews]
      summary: Submit feedback and verdict for interview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInterviewFeedbackRequest"
      responses:
        "200":
          description: Updated interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInterviewWithCandidate"

  # ── OFFERS ───────────────────────────────────────────────────────────────────
...
  /joining:
    get:
      operationId: listJoining
      tags: [joining]
      summary: List joining records
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Joining records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoiningWithCandidate"
    post:
      operationId: createJoining
      tags: [joining]
      summary: Create joining record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJoiningRequest"
      responses:
        "201":
          description: Created joining record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoiningWithCandidate"
  /joining/{id}:
    patch:
      operationId: updateJoining
      tags: [joining]
      summary: Update joining status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJoiningRequest"
      responses:
        "200":
          description: Updated joining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoiningWithCandidate"
  /joining/{id}/followup:
    post:
      operationId: createFollowup
      tags: [joining]
      summary: Log a post-joining follow-up
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowupRequest"
      responses:
        "201":
          description: Created followup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostJoiningFollowup"

  # ── DASHBOARD ────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications read
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Done
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # AUTH ──────────────────────────────────────────────────────────────────────
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        gstin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, isActive, createdAt]
    CreateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        website:
          type: string
        gstin:
          type: string
        address:
          type: string
      required: [name]
    CompanyContact:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        designation:
          type: string
          nullable: true
        isPrimary:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, isPrimary, createdAt]
    CreateCompanyContactRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        designation:
          type: string
        isPrimary:
          type: boolean
      required: [name]

    # REQUIREMENTS ──────────────────────────────────────────────────────────────
...
    Candidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        initials:
          type: string
          nullable: true
        email:
          type: string
        phone:
          type: string
          nullable: true
        currentRole:
          type: string
          nullable: true
        experienceYears:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        source:
          type: string
          enum: [referral, portal, social, internal, direct]
        currentCtc:
          type: integer
          nullable: true
        expectedCtc:
          type: integer
          nullable: true
        noticeDays:
          type: integer
          nullable: true
        summary:
          type: string
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, source, createdAt]
...
    ClientInterviewWithCandidate:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        mode:
          type: string
        round:
          type: string
        feedbackClient:
          type: string
          nullable: true
        feedbackRecruiter:
          type: string
          nullable: true
        verdict:
          type: string
          enum: [selected, on_hold, rejected, no_show]
          nullable: true
        conductedAt:
          type: string
          format: date-time
          nullable: true
        candidate:
          $ref: "#/components/schemas/Candidate"
        createdAt:
          type: string
          format: date-time
      required: [id, pipelineId, scheduledAt, mode, round, createdAt, candidate]
    CreateClientInterviewRequest:
      type: object
      properties:
        pipelineId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        mode:
          type: string
        round:
          type: string
        clientContactId:
          type: string
      required: [pipelineId, scheduledAt]
    ClientInterviewFeedbackRequest:
      type: object
      properties:
        feedbackClient:
          type: string
        feedbackRecruiter:
          type: string
        verdict:
          type: string
          enum: [selected, on_hold, rejected, no_show]
        conductedAt:
          type: string
          format: date-time
      required: [verdict]

    # OFFERS ────────────────────────────────────────────────────────────────────
...
    JoiningWithCandidate:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        joiningDate:
          type: string
          format: date-time
          nullable: true
        actualJoinedAt:
          type: string
          format: date-time
          nullable: true
        noticePeriodDays:
          type: integer
          nullable: true
        noticeStartDate:
          type: string
          format: date-time
          nullable: true
        bgvStatus:
          type: string
        docCollectionStatus:
          type: string
        laptopIssued:
          type: boolean
        idCardIssued:
          type: boolean
        candidate:
          $ref: "#/components/schemas/Candidate"
        createdAt:
          type: string
          format: date-time
      required: [id, pipelineId, bgvStatus, docCollectionStatus, laptopIssued, idCardIssued, createdAt, candidate]
    CreateJoiningRequest:
      type: object
      properties:
        pipelineId:
          type: string
        joiningDate:
          type: string
          format: date-time
        noticePeriodDays:
          type: integer
        noticeStartDate:
          type: string
          format: date-time
      required: [pipelineId]
    UpdateJoiningRequest:
      type: object
      properties:
        joiningDate:
          type: string
          format: date-time
        actualJoinedAt:
          type: string
          format: date-time
        bgvStatus:
          type: string
        docCollectionStatus:
          type: string
        laptopIssued:
          type: boolean
        idCardIssued:
          type: boolean
    PostJoiningFollowup:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        checkType:
          type: string
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        conductedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        retentionStatus:
          type: string
          enum: [retained, resigned, terminated, unknown]
        createdAt:
          type: string
          format: date-time
      required: [id, pipelineId, checkType, retentionStatus, createdAt]
    CreateFollowupRequest:
      type: object
      properties:
        checkType:
          type: string
          enum: ["30_day", "60_day", "90_day"]
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
        retentionStatus:
          type: string
          enum: [retained, resigned, terminated, unknown]
      required: [checkType]

    # DASHBOARD ─────────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
        recipientId:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        entityType:
          type: string
          nullable: true
        entityId:
          type: string
          nullable: true
        isRead:
          type: boolean
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, recipientId, type, title, body, isRead, createdAt]